src/app/requirements.txt
Pillow==2.2.1
simplejson==2.6.2
South==1.0.2
djangorestframework==2.3.14
django-braces==1.4.0
django-extensions==1.5.5
django-crispy-forms==1.5.2
dj-database-url==0.3.0
dj-static==0.0.6
static==0.4
uwsgi
MySQL-python==1.2.5